chore: dir 구조 변경
